stage2/requirements.txt
pymupdf==1.24.14
openai==1.59.7
python-dotenv==1.0.1
